+++ src/main/java/com/kangui/talentsharehub/domain/homework/service/HomeworkService.java
@@ -2,6 +2,7 @@ package com.kangui.talentsharehub.domain.homework.service;
 
 import com.kangui.talentsharehub.domain.course.entity.Course;
 import com.kangui.talentsharehub.domain.course.repository.course.CourseRepository;
+import com.kangui.talentsharehub.domain.course.repository.student.StudentRepository;
 import com.kangui.talentsharehub.domain.homework.dto.request.CreateHomeworkForm;
 import com.kangui.talentsharehub.domain.homework.dto.request.RequestUpdateHomework;
 import com.kangui.talentsharehub.domain.homework.dto.response.ResponseHomework;
@@ -11,18 +12,17 @@ import com.kangui.talentsharehub.global.exception.AppException;
 import com.kangui.talentsharehub.global.exception.ErrorCode;
 import com.kangui.talentsharehub.global.file.FileStore;
 import com.kangui.talentsharehub.global.file.UploadFile;
+import com.kangui.talentsharehub.global.login.resolver.dto.Principal;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.io.IOException;
-import java.net.MalformedURLException;
 import java.util.List;
 
 @Service
-@Transactional(readOnly = true)
+@Transactional
 @RequiredArgsConstructor
 @Slf4j
 public class HomeworkService {
@@ -31,38 +31,50 @@ public class HomeworkService {
     private String homeworkPath;
 
     private final HomeworkRepository homeworkRepository;
+    private final StudentRepository studentRepository;
     private final CourseRepository courseRepository;
     private final FileStore fileStore;
 
-    public List<ResponseHomework> getHomeworksByCourseId(Long courseId) {
-        Course course = courseRepository.findById(courseId)
-                .orElseThrow(() -> new AppException(ErrorCode.COURSE_NOT_FOUND, "존재 하지 않는 강의 입니다."));
+    @Transactional(readOnly = true)
+    public List<ResponseHomework> getHomeworksByCourseId(final Principal principal, final Long courseId) {
+        if(!courseRepository.existsById(courseId)) {
+            throw new AppException(ErrorCode.COURSE_NOT_FOUND, "존재 하지 않는 강의 입니다.");
+        }
 
-        List<Homework> homeworks = homeworkRepository.findByCourseIdWithAttachmentFile(course.getId());
+        if(courseRepository.existsByCourseIdAndUserId(courseId, principal.userId())
+                && studentRepository.existsByCourseIdAndUserId(courseId, principal.userId())) {
+            throw new AppException(ErrorCode.FORBIDDEN, "강의 관계자만 과제 조회가 가능합니다.");
+        }
+
+        List<Homework> homeworks = homeworkRepository.findByCourseIdWithAttachmentFile(courseId);
 
         return homeworks.stream()
-                .map(ResponseHomework::new)
+                .map(ResponseHomework::of)
                 .toList();
     }
 
-    @Transactional
-    public Long createHomework(CreateHomeworkForm createHomeworkForm) {
-        if (!createHomeworkForm.getStartDate().isAfter(createHomeworkForm.getEndDate())) {
-            throw new AppException(ErrorCode.INVALID_DATE_RANGE, "종료 일이 시작 일보다 빠를 수 없습니다.");
-        }
-
-        Course course = courseRepository.findById(createHomeworkForm.getCourseId())
+    public Long createHomework(
+            final Principal principal,
+            final CreateHomeworkForm createHomeworkForm,
+            final Long courseId
+    ) {
+        Course course = courseRepository.findByIdWithUser(courseId)
                 .orElseThrow(() -> new AppException(ErrorCode.COURSE_NOT_FOUND, "존재 하지 않는 강의입니다."));
 
-        List<UploadFile> uploadFiles = null;
-
-        try {
-            uploadFiles = fileStore.storeFiles(createHomeworkForm.getAttachmentFiles(), homeworkPath);
-        } catch (IOException e) {
-            throw new AppException(ErrorCode.FILE_UPLOAD_FAILED, "파일 업로드에 실패했습니다.");
+        if(course.getUser().getId().equals(principal.userId())) {
+            throw new AppException(ErrorCode.FORBIDDEN, "선생님만 과제 생성이 가능합니다.");
         }
 
-        Homework homework = createHomeworkForm.toEntity(course);
+        List<UploadFile> uploadFiles = fileStore.storeFiles(createHomeworkForm.getAttachmentFiles(), homeworkPath);
+
+        Homework homework = new Homework(
+                course,
+                createHomeworkForm.getTitle(),
+                createHomeworkForm.getContents(),
+                null,
+                createHomeworkForm.getStartDate(),
+                createHomeworkForm.getEndDate()
+        );
 
         uploadFiles.stream()
                 .map(UploadFile::toHomeworkAttachmentFile)
@@ -71,35 +83,34 @@ public class HomeworkService {
         return homeworkRepository.save(homework).getId();
     }
 
-    @Transactional
-    public Long updateHomework(Long homeworkId, RequestUpdateHomework requestUpdateHomework) {
-        if (!requestUpdateHomework.getStartDate().isAfter(requestUpdateHomework.getEndDate())) {
-            throw new AppException(ErrorCode.INVALID_DATE_RANGE, "종료 일이 시작 일보다 빠를 수 없습니다.");
+    public Long updateHomework(
+            final Principal principal,
+            final Long homeworkId,
+            final RequestUpdateHomework requestUpdateHomework,
+            final Long courseId
+    ) {
+        if(courseRepository.existsByCourseIdAndUserId(courseId, principal.userId())) {
+            throw new AppException(ErrorCode.FORBIDDEN, "선생님만 과제 수정이 가능합니다.");
         }
-        
+
         Homework homework = homeworkRepository.findById(homeworkId)
                 .orElseThrow(() -> new AppException(ErrorCode.HOMEWORK_NOT_FOUND, "과제가 존재하지 않습니다."));
 
-        homework.updateHomework(
-                requestUpdateHomework.getTitle(),
-                requestUpdateHomework.getContents(),
-                requestUpdateHomework.getStartDate(),
-                requestUpdateHomework.getEndDate());
+        homework.updateHomework(requestUpdateHomework);
 
         return homework.getId();
     }
 
-    @Transactional
-    public void deleteHomework(Long homeworkId) {
-        Homework homework = homeworkRepository.findByIdWithAttachmentFile(homeworkId)
+    public void deleteHomework(final Principal principal, final Long homeworkId, final Long courseId) {
+        if(courseRepository.existsByCourseIdAndUserId(courseId, principal.userId())) {
+            throw new AppException(ErrorCode.FORBIDDEN, "선생님만 과제 수정이 가능합니다.");
+        }
+
+        final Homework homework = homeworkRepository.findByIdWithAttachmentFile(homeworkId)
                 .orElseThrow(() -> new AppException(ErrorCode.HOMEWORK_NOT_FOUND, "과제가 존재하지 않습니다."));
 
         homework.getHomeworkAttachmentFile().forEach(homeworkAttachmentFile -> {
-            try {
-                fileStore.deleteFile(homeworkAttachmentFile.getStoreFileName(), homeworkPath);
-            } catch (MalformedURLException e) {
-                throw new AppException(ErrorCode.FILE_DELETE_FAILED, "과제 첨부 파일 삭제에 실패했습니다.");
-            }
+            fileStore.deleteFile(homeworkAttachmentFile.getStoreFileName(), homeworkPath);
         });
 
         homeworkRepository.delete(homework);
